bundle: drop fixed `:latest` version from controller image-name
The controller image-name should use the `TAG` just as other images.
When this is not done, the bundle includes the following reference to
the image, and fails to deploy:

    quay.io/csiaddons/k8s-controller:latest:latest`

By using `TAG` as the other container images do, the bundle will refer
to the image with `:latest`, or the `TAG` that was used for creating a
release.

Signed-off-by: Niels de Vos <ndevos@redhat.com>