build: version the bundle/CSV the same as the tag of the container-images
When a release is made, the git-tag is used for the version of the
container-images. It seems that ODF-Operator assumes the tag of the
container-image matches what is used as a version of the bundle/CSV.

There is no hard requirement that these versions are the same, it is
just an assumption of ODF-Operator. However, it is the first operator
that deploys CSI-Addons, and changing the operator seems more complex
than versioning csi-addons differently.

The default version in csi-addons.clusterserviceversion.yaml has now
been set to 0.1.1, the latest git-tag that was published.

Fixes: #82
Signed-off-by: Niels de Vos <ndevos@redhat.com>