build: generate a clusterserviceversion manifest with PACKAGE_NAME
When PACKAGE_NAME is set while building the bundle, the base for the
clusterserviceversion manifest is not detected correctly. When that
happens, operator-sdk generates a new clusterserviceversion manifest
with default values.

Because the values in the clusterserviceversion manifest have a meaning,
and were tested with different deployments, a default new manifest is
not what we want included in the bundle.

By replacing the @PACKAGE_NAME@ placeholder in the
clusterserviceversion.yaml.in file before generating the bundle, the
resulting bundle now contains all the values from the base manifest,
with the correct package name.

This can be verified by running commands like:

  - make bundle PACKAGE_NAME=hello-world
  - make bundle PACKAGE_NAME=hello-world TAG=v0.0.1-alpha.42

Signed-off-by: Niels de Vos <ndevos@redhat.com>