vendor: bump github.com/operator-framework/operator-sdk from 1.17.0 to 1.19.1
This updates operator-sdk to v1.19.1, the update to v1.19.0 failed
earlier and was reverted. With this vendor update, everything seems to
build fine still.

There is an update for the github.com/containerd/containerd package, but
that does not work with github.com/deislabs/oras@v0.11.1. This is a
known issue, the workaround comes from the operator-sdk go.mod file.

Signed-off-by: Niels de Vos <ndevos@redhat.com>