NetworkFence: Explicitly mentioned Fenced and Unfenced state
Currently, the assumption is that, the resource by default
strives to ensure that the CIDRs mentioned in the spec are
fenced. Deletion of the resource makes those CIDRs to have
been unfenced.

This has some issues when fencing is consumed for operations
such as Disaster Recovery. Hence, changed the CR to have
explicit state as either fence or unfence.

Co-authored-by: Raghavendra Talur <raghavendra.talur@gmail.com>
Signed-off-by: Raghavendra M <raghavendra@redhat.com>
(cherry picked from commit 848e8b27d3950f315360b9fea0e036cd59998cdb)