Bug 2233071: rebase: bump k8s.io dependencies to v0.26.7
Bumping the k8s.io dependencies to v0.26.7. The main intention
is to update the client-go library dependency for kube 1.27+,
where aggregated discovery is enabled. With client-go lower than
v0.26.4 in some cases it is possible that the discovery response
comes with the malformed response with nil GVK leading to Crash
Loop Back Off state.

Upstream kubernetes issue:
kubernetes/kubernetes#116603

Signed-off-by: karthik-us <ksubrahm@redhat.com>