use NormalizeLeaseName to store the key
namespace + "/" + leader identity(pod name)
is the key for the connection.
this key is used by GetLeaderByDriver to
get the connection.
csi-lib-utils/leaderelection:sanitizeName() is
used to sanitize the leader identity used
for the leases csiaddonsnode need to store the
key with same format so that it can be used
to get the connection.

Signed-off-by: Madhu Rajanna <madhupr007@gmail.com>
(cherry picked from commit 6c312483aae0bfe47f4a8ce7a19ee11ec6be912d)